+++ Tests/GRDBTests/ValueObservationCompactMapTests.swift
@@ -39,10 +39,13 @@ class ValueObservationCompactMapTests: GRDBTestCase {
             }
             
             // Start observation
-            let observer = try observation.start(in: dbWriter) { count in
-                counts.append(count)
-                notificationExpectation.fulfill()
-            }
+            let observer = observation.start(
+                in: dbWriter,
+                onError: { error in XCTFail("Unexpected error: \(error)") },
+                onChange: { count in
+                    counts.append(count)
+                    notificationExpectation.fulfill()
+            })
             try withExtendedLifetime(observer) {
                 try dbWriter.writeWithoutTransaction { db in
                     try db.execute(sql: "INSERT INTO t DEFAULT VALUES")