+++ Tests/GRDBTests/ValueObservationRecordTests.swift
@@ -34,10 +34,13 @@ class ValueObservationRecordTests: GRDBTestCase {
         notificationExpectation.expectedFulfillmentCount = 4
         
         let observation = SQLRequest<Player>(sql: "SELECT * FROM t ORDER BY id").observationForAll()
-        let observer = try observation.start(in: dbQueue) { players in
-            results.append(players)
-            notificationExpectation.fulfill()
-        }
+        let observer = observation.start(
+            in: dbQueue,
+            onError: { error in XCTFail("Unexpected error: \(error)") },
+            onChange: { players in
+                results.append(players)
+                notificationExpectation.fulfill()
+        })
         try withExtendedLifetime(observer) {
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "INSERT INTO t (id, name) VALUES (1, 'foo')") // +1
@@ -70,10 +73,13 @@ class ValueObservationRecordTests: GRDBTestCase {
         notificationExpectation.expectedFulfillmentCount = 4
         
         let observation = Player.observationForAll()
-        let observer = try observation.start(in: dbQueue) { players in
-            results.append(players)
-            notificationExpectation.fulfill()
-        }
+        let observer = observation.start(
+            in: dbQueue,
+            onError: { error in XCTFail("Unexpected error: \(error)") },
+            onChange: { players in
+                results.append(players)
+                notificationExpectation.fulfill()
+        })
         try withExtendedLifetime(observer) {
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "INSERT INTO t (id, name) VALUES (1, 'foo')") // +1
@@ -106,10 +112,13 @@ class ValueObservationRecordTests: GRDBTestCase {
         notificationExpectation.expectedFulfillmentCount = 4
         
         let observation = SQLRequest<Player>(sql: "SELECT * FROM t ORDER BY id DESC").observationForFirst()
-        let observer = try observation.start(in: dbQueue) { player in
-            results.append(player)
-            notificationExpectation.fulfill()
-        }
+        let observer = observation.start(
+            in: dbQueue,
+            onError: { error in XCTFail("Unexpected error: \(error)") },
+            onChange: { player in
+                results.append(player)
+                notificationExpectation.fulfill()
+        })
         try withExtendedLifetime(observer) {
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "INSERT INTO t (id, name) VALUES (1, 'foo')") // +1